ToolchainDefinition: add arm-c4ev3-linux-uclibceabi
We had so far multiple users inquire about support for newer C++ standard while remaining
compatible with stock firmware, i.e. same old glibc and kernel. Add the uclibc configuration
of the C4EV3.Toolchain to address this. For now leave it at the bottom of the list, as binary
releases are only available for Linux. When it's part of the Windows Installer, it should be
move to the top.